+++ src/Sentry.php
@@ -1,26 +1,13 @@
 <?php
-/**
- * Sentry plugin for Craft CMS 3.x
- *
- * Error tracking that helps developers monitor and fix crashes in real time. Iterate continuously. Boost efficiency. Improve user experience.
- *
- * @link      https://github.com/lukeyouell
- * @copyright Copyright (c) 2017 Luke Youell
- */
-
 namespace lukeyouell\sentry;
-
-use lukeyouell\sentry\services\SentryService;
-use lukeyouell\sentry\models\Settings;
-use lukeyouell\sentry\variables\SentryVariable;
 
 use Craft;
 use craft\base\Plugin;
 use craft\events\ExceptionEvent;
-use craft\events\RegisterUrlRulesEvent;
 use craft\web\ErrorHandler;
-use craft\web\UrlManager;
-use craft\web\twig\variables\CraftVariable;
+
+use lukeyouell\sentry\models\Settings;
+use lukeyouell\sentry\services\SentryService;
 
 use yii\base\Event;
 
@@ -39,30 +26,15 @@ class Sentry extends Plugin
         parent::init();
         self::$plugin = $this;
 
-        Event::on(
-            UrlManager::class,
-            UrlManager::EVENT_REGISTER_CP_URL_RULES,
-            function (RegisterUrlRulesEvent $event) {
-                $event->rules['settings/sentry/authentication'] = 'sentry/settings/authentication';
-                $event->rules['settings/sentry/project'] = 'sentry/settings/project';
-                $event->rules['settings/sentry/excluded-codes'] = 'sentry/settings/excluded-codes';
-            }
-        );
-
-        Event::on(
-            CraftVariable::class,
-            CraftVariable::EVENT_INIT,
-            function (Event $event) {
-                $variable = $event->sender;
-                $variable->set('Sentry', SentryVariable::class);
-            }
-        );
+        $this->setComponents([
+            'sentry' => SentryService::class,
+        ]);
 
         Event::on(
             ErrorHandler::className(),
             ErrorHandler::EVENT_BEFORE_HANDLE_EXCEPTION,
             function(ExceptionEvent $event) {
-                SentryService::handleException($event->exception);
+                $this->sentry->handleException($event->exception);
             }
         );
     }
@@ -77,6 +49,17 @@ class Sentry extends Plugin
 
     protected function settingsHtml(): string
     {
-        return Craft::$app->view->renderTemplate('sentry/settings');
+        $settings = $this->getSettings();
+        $settings->validate();
+        $overrides = Craft::$app->getConfig()->getConfigFromFile(strtolower($this->handle));
+
+        return Craft::$app->view->renderTemplate('sentry/settings',
+            [
+                'plugin'    => $this,
+                'title'     => $this->handle,
+                'settings'  => $settings,
+                'overrides' => array_keys($overrides)
+            ]
+        );
     }
 }